Feat: 컬렉션뷰 분리
