+++ theLastPush/theLastPush/Screens/Main/MainVC.swift
@@ -5,12 +5,13 @@
 //  Created by 지영 on 5/19/25.
 //
 
-import SnapKit
 import UIKit
+import SnapKit
 
 class MainVC: UIViewController {
     
-    private let members: [Member] = memberData
+//    private let members: [Member] = memberData
+    private let collectionView = MemberCollectionView()
     
     private let titleLabel: UILabel = {
         let label = UILabel()
@@ -32,21 +33,7 @@ class MainVC: UIViewController {
         return button
     }()
     
-    private lazy var membersCollectionView: UICollectionView = {
-        let layout = UICollectionViewFlowLayout()
-        layout.scrollDirection = .vertical
-        layout.minimumLineSpacing = 20
-        layout.minimumInteritemSpacing = 20
-        
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collectionView.backgroundColor = .clear
-        collectionView.showsVerticalScrollIndicator = false
-        collectionView.register(MemberCell.self, forCellWithReuseIdentifier: MemberCell.identifier)
-        collectionView.delegate = self
-        collectionView.dataSource = self
-        return collectionView
-    }()
-
+    
 	override func viewWillAppear(_ animated: Bool) {
 		navigationController?.isNavigationBarHidden = true
 	}
@@ -59,14 +46,18 @@ class MainVC: UIViewController {
         super.viewDidLoad()
         setupView()
         setupConstraints()
-        
-        view.backgroundColor = .white
     }
     
     private func setupView()  {
         view.addSubview(titleLabel)
         view.addSubview(introButton)
-        view.addSubview(membersCollectionView)
+        view.addSubview(collectionView)
+        
+        view.backgroundColor = .white
+        
+        collectionView.onMemberSelected = { [weak self] member in
+            self?.handleMemberSelection(member)
+        }
     }
     
     private func setupConstraints() {
@@ -83,16 +74,13 @@ class MainVC: UIViewController {
             $0.height.equalTo(60)
         }
         
-        introButton.addTarget(self, action: #selector(introButtonTapped), for: .touchUpInside)
-    }
-    
-    
-    private func setUpMembersCollectionView() {
-        membersCollectionView.snp.makeConstraints {
+        collectionView.snp.makeConstraints {
             $0.top.equalTo(introButton.snp.bottom).offset(60)
             $0.leading.trailing.equalToSuperview().inset(20)
             $0.bottom.equalTo(view.safeAreaLayoutGuide.snp.bottom).offset(-8)
         }
+        
+        introButton.addTarget(self, action: #selector(introButtonTapped), for: .touchUpInside)
     }
     
     @objc private func introButtonTapped() {
@@ -100,37 +88,14 @@ class MainVC: UIViewController {
         vc.modalPresentationStyle = .fullScreen
         present(vc, animated: true)
     }
-}
-
-extension MainVC: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return members.count
-    }
     
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: MemberCell.identifier, for: indexPath) as? MemberCell else { return UICollectionViewCell() }
-        cell.configure(with: members[indexPath.item])
-        return cell
-    }
-    
-   
-}
-
-extension MainVC: UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let member = members[indexPath.item]
+    private func handleMemberSelection(_ member: Member) {
         var nextViewController: UIViewController
-        print("\(member.name) 페이지로 이동!")
-
         switch member.name {
         case "조성준":
             nextViewController = JuseojoVC()
         case "양지영":
-            guard let milou = members.first(where: { $0.name == "양지영" }) else {
-                    print("못찾음")
-                    return
-                }
-            nextViewController = MilouVC(member: milou)
+            nextViewController = MilouVC(member: member)
         case "서광용":
             nextViewController = MoriVC()
         case "이태윤":
@@ -141,13 +106,11 @@ extension MainVC: UICollectionViewDelegate {
         
         self.navigationController?.pushViewController(nextViewController, animated: true)
     }
+    
 }
 
-extension MainVC: UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        let width = (collectionView.frame.width - 20) / 2
-        return CGSize(width: width, height: width * 1.3)
-    }
-}
+
+
+
